+++ lib/src/material_options.dart
@@ -9,6 +9,10 @@ class MaterialOptions {
   final bool useDetailsView;
   final String selectCircleStrokeColor;
   final String selectionLimitReachedText;
+  final String textOnNothingSelected;
+  final String backButtonDrawable;
+  final String okButtonDrawable;
+  final bool autoCloseOnSelectionLimit;
 
   const MaterialOptions({
     this.actionBarColor,
@@ -21,6 +25,10 @@ class MaterialOptions {
     this.useDetailsView,
     this.selectCircleStrokeColor,
     this.selectionLimitReachedText,
+    this.textOnNothingSelected,
+    this.backButtonDrawable,
+    this.okButtonDrawable,
+    this.autoCloseOnSelectionLimit,
   });
 
   Map<String, String> toJson() {
@@ -35,6 +43,11 @@ class MaterialOptions {
       "useDetailsView": useDetailsView == true ? "true" : "false",
       "selectCircleStrokeColor": selectCircleStrokeColor ?? "",
       "selectionLimitReachedText": selectionLimitReachedText ?? "",
+      "textOnNothingSelected": textOnNothingSelected ?? "",
+      "backButtonDrawable": backButtonDrawable ?? "",
+      "okButtonDrawable": okButtonDrawable ?? "",
+      "autoCloseOnSelectionLimit":
+          autoCloseOnSelectionLimit == true ? "true" : "false"
     };
   }
 }